% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pileup_readid}
\alias{pileup_readid}
\title{list alleles and the read id that supports them at a given pos}
\usage{
pileup_readid(bam_path, seqname, pos)
}
\description{
list alleles and the read id that supports them at a given pos
}
