% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{variant_matrix}
\alias{variant_matrix}
\title{produce a matrix of variants and cell barcodes}
\usage{
variant_matrix(bam_path, seqname, pos, barcodes)
}
\description{
produce a matrix of variants and cell barcodes
}
